+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.10.29
+
+### Patch Changes
+
+- Updated dependencies [c9fd939]
+    - @sap-ux/store@1.5.0
+
 ## 0.10.28
 
 ### Patch Changes