+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.40.11
+
+### Patch Changes
+
+- Updated dependencies [be67fc4]
+    - @sap-ux/project-access@1.34.2
+    - @sap-ux/fiori-annotation-api@0.9.3
+
+## 0.40.10
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/fiori-annotation-api@0.9.2
+
+## 0.40.9
+
+### Patch Changes
+
+- Updated dependencies [4d6695f]
+    - @sap-ux/fiori-annotation-api@0.9.1
+
 ## 0.40.8
 
 ### Patch Changes