+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.9.3
+
+### Patch Changes
+
+- Updated dependencies [be67fc4]
+    - @sap-ux/project-access@1.34.2
+    - @sap-ux/cds-odata-annotation-converter@0.7.1
+
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/cds-odata-annotation-converter@0.7.1
+
+## 0.9.1
+
+### Patch Changes
+
+- 4d6695f: fix: add missing JSDoc
+- Updated dependencies [4d6695f]
+    - @sap-ux/cds-odata-annotation-converter@0.7.1
+
 ## 0.9.0
 
 ### Minor Changes