+++ packages/launch-config/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/launch-config
 
+## 0.10.51
+
+### Patch Changes
+
+- Updated dependencies [be67fc4]
+- Updated dependencies [be67fc4]
+    - @sap-ux/ui5-config@0.29.13
+    - @sap-ux/project-access@1.34.2
+
+## 0.10.50
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+
 ## 0.10.49
 
 ### Patch Changes