+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.6.45
+
+### Patch Changes
+
+- @sap-ux/project-input-validator@0.6.45
+- @sap-ux/inquirer-common@0.10.13
+
+## 0.6.44
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.12
+
+## 0.6.43
+
+### Patch Changes
+
+- @sap-ux/project-input-validator@0.6.44
+- @sap-ux/inquirer-common@0.10.11
+
 ## 0.6.42
 
 ### Patch Changes